refactor(ui): consolidate studio endpoints with generic core APIs
