+++ app/ui/static/js/studio.js
@@ -346,7 +346,7 @@ document.addEventListener('keydown', e => {
 });
 
 // ── Interactive Pipeline Actions ────────────────────────────────
-async function postUI(path, body = {}) {
+async function postAPI(path, body = {}) {
   const res = await (window.authFetch || fetch)(path, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
@@ -363,7 +363,7 @@ window.approveTalk = async function(id) {
   const notes = (document.getElementById('review-notes-input') || {}).value || '';
   const btn = document.getElementById('btn-approve');
   const originalHtml = btn ? btn.innerHTML : '';
-  if (btn) { btn.disabled = true; btn.innerHTML = '<span class="spinner spinner-sm"></span> Processing Pipeline...'; }
+  if (btn) { btn.disabled = true; btn.innerHTML = '<span class="spinner spinner-sm"></span> Processing...'; }
 
   const progressWrap = document.getElementById('pipeline-progress-wrap');
   const progressFill = document.getElementById('pipeline-progress-fill');
@@ -379,34 +379,71 @@ window.approveTalk = async function(id) {
   if (progressDesc) progressDesc.textContent = 'Executing processing pipeline...';
 
   try {
-    await postUI(`/studio/talks/${id}/approve`, {
-      decision: 'approved',
-      note: notes || 'Approved in review studio',
-      start_sec: inPointSec,
-      end_sec: outPointSec,
-    });
+    const talkStatus = typeof TALK_STATUS !== 'undefined' ? TALK_STATUS : '';
+    if (talkStatus === 'pending_approval') {
+      await postAPI(`/talks/${id}/approve`, { decision: 'approve' });
+    } else if (talkStatus === 'pending_bounds' || talkStatus === 'needs_work' || talkStatus === 'waiting_for_files') {
+      const cutStart = formatTimecode(inPointSec).slice(0, 8);
+      const cutEnd = formatTimecode(outPointSec).slice(0, 8);
+      await postAPI(`/talks/${id}/cut`, { cut_start: cutStart, cut_end: cutEnd });
+    } else if (talkStatus === 'preview') {
+      await postAPI(`/talks/${id}/review`, { decision: 'approve', note: notes || 'Approved in review studio' });
+    } else if (talkStatus === 'pending_intro_outro') {
+      const includeIntro = document.getElementById('check-include-intro') ? document.getElementById('check-include-intro').checked : true;
+      const includeOutro = document.getElementById('check-include-outro') ? document.getElementById('check-include-outro').checked : true;
+      await postAPI(`/talks/${id}/assemble`, {
+        include_intro: includeIntro,
+        include_outro: includeOutro,
+        intro_source: 'generated',
+        outro_source: 'generated',
+      });
+    } else {
+      await postAPI(`/talks/${id}/approve`, { decision: 'approve' });
+    }
     if (progressFill) progressFill.style.width = '100%';
     if (progressPct) progressPct.textContent = '100%';
     if (progressDesc) progressDesc.textContent = 'Pipeline complete! Reloading studio...';
     setTimeout(() => location.reload(), 400);
   } catch (err) {
-    alert(`Pipeline execution failed: ${err.message}`);
+    alert(`Pipeline action failed: ${err.message}`);
     if (progressWrap) progressWrap.style.display = 'none';
     if (btn) { btn.disabled = false; btn.innerHTML = originalHtml; }
   }
 };
 
+
 window.rejectTalk = async function(id) {
   const notes = (document.getElementById('review-notes-input') || {}).value || '';
-  if (!confirm('Reject this talk bounds?')) return;
+  if (!confirm('Reject this talk?')) return;
   const btn = document.getElementById('btn-reject');
   const originalHtml = btn ? btn.innerHTML : '';
   if (btn) { btn.disabled = true; btn.innerHTML = '<span class="spinner spinner-sm"></span> Rejecting...'; }
   try {
-    await postUI(`/studio/talks/${id}/reject`, { decision: 'rejected', note: notes || 'Rejected in review studio' });
+    const talkStatus = typeof TALK_STATUS !== 'undefined' ? TALK_STATUS : '';
+    if (talkStatus === 'pending_approval') {
+      await postAPI(`/talks/${id}/approve`, { decision: 'reject' });
+    } else if (talkStatus === 'preview') {
+      await postAPI(`/talks/${id}/review`, { decision: 'reject', note: notes || 'Rejected in review studio' });
+    } else {
+      await postAPI(`/talks/${id}/abort`);
+    }
     location.reload();
   } catch (err) {
     alert(`Rejection failed: ${err.message}`);
+    if (btn) { btn.disabled = false; btn.innerHTML = originalHtml; }
+  }
+};
+
+window.requestChangesTalk = async function(id) {
+  const notes = (document.getElementById('review-notes-input') || {}).value || '';
+  const btn = document.getElementById('btn-needs-work');
+  const originalHtml = btn ? btn.innerHTML : '';
+  if (btn) { btn.disabled = true; btn.innerHTML = '<span class="spinner spinner-sm"></span> Requesting changes...'; }
+  try {
+    await postAPI(`/talks/${id}/review`, { decision: 'needs_work', note: notes || 'Needs work' });
+    location.reload();
+  } catch (err) {
+    alert(`Request changes failed: ${err.message}`);
     if (btn) { btn.disabled = false; btn.innerHTML = originalHtml; }
   }
 };
@@ -416,20 +453,11 @@ window.retryTalk = async function(id) {
   const originalHtml = btn ? btn.innerHTML : '';
   if (btn) { btn.disabled = true; btn.innerHTML = '<span class="spinner spinner-sm"></span> Resetting...'; }
   try {
-    await postUI(`/studio/talks/${id}/retry`);
+    await postAPI(`/talks/${id}/abort`);
     location.reload();
   } catch (err) {
-    alert(`Retry failed: ${err.message}`);
+    alert(`Reset failed: ${err.message}`);
     if (btn) { btn.disabled = false; btn.innerHTML = originalHtml; }
-  }
-};
-
-window.setTalkStatus = async function(id, newStatus) {
-  try {
-    await postUI(`/studio/talks/${id}/status`, { status: newStatus });
-    location.reload();
-  } catch (err) {
-    alert(`Status update failed: ${err.message}`);
   }
 };
 
@@ -446,7 +474,7 @@ window.handleVideoFileUpload = async function(e, talkId) {
     const fd = new FormData();
     fd.append('file', file);
 
-    const res = await (window.authFetch || fetch)(`/studio/talks/${talkId}/upload-recording`, {
+    const res = await (window.authFetch || fetch)(`/talks/${talkId}/upload`, {
       method: 'POST',
       body: fd,
     });
@@ -456,7 +484,7 @@ window.handleVideoFileUpload = async function(e, talkId) {
       throw new Error(err.detail || `Upload failed with status ${res.status}`);
     }
 
-    const data = await res.json();
+    await res.json();
     location.reload();
   } catch (err) {
     alert(`Video upload failed: ${err.message}`);
